Bulletin board/BulletinBoard/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==39.0.1
defusedxml==0.7.1
Django==4.1.7
Django-Accounts==0.1
django-allauth==0.52.0
django-bootstrap4==22.3
django-braces==1.15.0
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-filter==22.1
django-js-asset==2.0.0
dnspython==2.3.0
eventlet==0.33.3
greenlet==2.0.2
idna==3.4
kombu==5.2.4
oauthlib==3.2.2
prompt-toolkit==3.0.38
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
